Add scripts to get Change-Ids from parent commits
The newly added scripts make it possible to get information about
commits that depend on not yet merged parent commits of the current
project.

An example should make things clearer. Given there's currently code
review going on for three geocouch changes A, B and C. Change B
depends on a couchdb change named X. Now we run commit validation for
all geocouch changes:

 - A: the validation passes as it doesn't depend on anything
 - B: the validation passes thanks to the already existing script,
   that will also checkout change X from couchdb as it has they share
   the same Change-Id
 - C: without the new scripts, the commit validation would fail, as
   there is no couchdb change that has the same Change-Id as change C
   has. With the scripts the validation passes, as it checks all the
   parent commits of change C. Hence it finds change B and checks if
   any other projects have a commit with the same Change-Id. Commit X
   is found and the corresponding change will be fetched.

The search for such commits is restricted to not yet merged commits,
as merged commit will automatically be checked out by repo.

Possible future improvement: Also check dependencies that are not
directly tied to the project of the current commit. An example would
be a parent commit of X in couchdb named W. If that one depends on
change M in couchstore we are currently not able to detect that.

As such complex dependencies happen rarely and there's workaround with
waiting for the validation of change C until W and M are merged. I
think that the current approach is sufficient, hence a valid trade-off
to keep things simple.

Change-Id: Ieefb0a794b5a8bed826e3c8e9d7a17937f4cb7ad